Initial commit
This is the code from Chapter 4: Scanning of the Java tree-walk
interpreter. I don't really know Java so the code is pretty much
verbatim and the project structure (or package structure?) somewhat
mimics the authors'.

I also am a bit clueless when it comes to building Java projects "by
hand" so I can slap together a nice Makefile or ninja.build. I'm
definitely not looking to bring in the standard Java build frameworks
for something _academic_ like this.

I _might_ (but probably won't) come back with an auto-formatter and/or
linter to make the code a little bit more idiomatic or nicer to look
at... after all, I'm not dealing with constraints of fitting code in a
book so I can get rid of single-line loops *shudders*.