Refactor generate_ast to build up file in memory
I wasn't too fond of all the `f.write(XXX\n)` statements of before, so
I now put the file together by shoving every line in a list and joining
them in the end. The source files will never be big enough so that this
is a problem.

Everything else is still pretty 1:1 with the canonical Java version.