Add Ninja build file
I decided to go with Ninja because it's something I've used before and
Make still looks like magic to me.
It's not the best because it seems Java _needs_ to compile everything in
one go, so there's really no incremental builds. Also, every class and
subclass gets a separate file (??) so it's kind of a pain to write by
hand (which to be fair, is not the correct way to use Ninja).

Still, it works. And beats copying the build commands from the README
every time. $.$