Add static resolution
This is the code from Chapter 11: Resolving and Binding.

This one is a bit of a doozy - the concepts are fairly abstract and the
Interpreter and Resolver classes are tightly coupled. After typing all
of the code I had to turn back and do a double-take to try and
consolidate what I had just created.

Deleted the `challenge.lox` from the Chapter 8 examples as the
interpreter now correctly spits out an error on it. I wasn't aware of
Lox's rules on variable scoping when I included it.